requirements.txt
flask
requests
pdf2image
pytesseract
spacy
torch
sentence-transformers
rank-bm25
numpy
python-dotenv